requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
mediapipe==0.10.10  # Updated to available version
ultralytics==8.0.0
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
matplotlib==3.7.1
plotly==5.14.1
pyyaml==6.0
tqdm==4.65.0